Fix virtualenv python path
The virtualenv was loaded first, then the PATH was changed. Therefore
the python used in the startup script may not be the one from the
virtualenv. This caused problems with other releases that manage other
python installs.

This change moves the activation of the virtualenv after the PATH change
to avoid this issue.