Move requirements*.txt to src/ instead of blobs
We want to be able to track and change the files for requirements.txt
and requirements-opt.txt, which are currently managed as external blobs.

This is required to enable different datadog plugins which require
additional python dependencies. Having this file in code will be make it
easier to fork the release and do the required changes, and maintain
changes via PRs.