Add option to set hostname with bosh name.index
Setting `use_bosh_hostname` will change the hostname as it should be
displayed in datadog as $jobname.$index, from the spec struct.

If for some reason they are unknown (e.g. bosh-init) they will be set
to `unknown`. In that case the user should use `hostname`.

The property `hostname` still takes precedence.

We also add the tag `bosh-id` when `include_bosh_tags` is set to be able
to uniquely identify the instance.

We cannot get the real hostname (bosh-agent-id) in the template generation
due limitations in bosh.