Enable 3rd party release custom checks
Add logic that collects custom datadog checks from 3rd party releases.
Scripts placed in

${JOB_PATH}/config/datadog-integrations/${checkname}.py

will be copied to checks.d directory of the datadog agent. Configuration for
these checks can be done using already built in 3rd party configuration
merging functionality. Simply - put the config in the same directory and
ensure it has `.yaml` suffix.

This functionality will break datadog-agent job startup if it finds two
custom checks with the same name. This is so that it doesn't overwrite
checks and doesn't fail silently when it's not clear which check should be
used.