Prepend deployment name to hostname
If reporting from multiple deployments that happen to use the same job
names (the likelihood of this is very high), Datadog will get confused.

For example, we have 2 Concourse deployments. Whichever instance starts
the datadog-agent jobs last will be considered the "real" e.g. worker.0
or db.0 and the other deployment will "refuse" to report metrics for that
instance name.