Fix use_bosh_hostname logic
if_p only checks if property exists (is set). Which is always true for
use_bosh_hostname as we give it default of false in the spec.

This means that template would _always_ render as if use_bosh_hostname was
true, regardless of actual value. Change it to real `if` to fix logic. We
don't need to test if the property is set at all, because we give it default
value.