Bundle and build python 2.7 in the release
We want to be able to use this datadog-agent in the new Xenial
stemcells. But we are currently having issues because
Xenial stemcells do not include python binaries.

The agent fails with errors like:

   Could not find platform dependent libraries <exec_prefix>
   Consider setting $PYTHONHOME to <prefix>[:<exec_prefix>]
   Traceback (most recent call last):
   File "/var/vcap/jobs/datadog-agent/packages/datadog-agent/agent/agent.py", line 13, in <module>
      from config import get_version, initialize_logging  # noqa
   File "/var/vcap/data/packages/datadog-agent/8476309a9054a2b637422d241a882303d1ad56be/agent/config.py", line 6, in <module>
      import ConfigParser
   ImportError: No module named ConfigParser

This release relies on:
 - python being isntalled in the stemcell
 - be a ubuntu based stemcell
 - be able to install python-dev packages during compilation of
   the bosh package `datadog-agent`.

In this commit we change the approach to follow the recommendation
for bosh releases: bundle all the source code of the dependencies
as blobs and build all at compile time.

To do so we will:

 - Remove the old python-dev
 - Add a new package python-2.7, including pip (based on [1], but updated)
 - Update the datadog-agent to:
   - use the new python-2.7 package for packaging and for execution.
   - bundle ALL the python dependencies and compile dependencies
   - compile and install the dependencies

We get all the datadog agent dependencies using `pip download`,
and other libraries were downloaded adhoc. `pip` receives the option
`--find-links` to search for these files later.

The helper script ./scr/datadog-agent/download-requirements.sh
automates the download of the dependencies.

We had to use `pkg-config` because the pip packages require it
to find the build libraries like libcurl or libffi.

We had to update pyro4 dependency because the version 4.35 was not
available for download.

[1] https://github.com/bosh-packages/python-release